% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxquant_decoy.R
\docType{data}
\name{maxquant_decoy}
\alias{maxquant_decoy}
\title{Decoy search results from MaxQuant on trumpp}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(maxquant_decoy)
}
\description{
Decoy search results from MaxQuant on trumpp
}
\references{
Raffel, S., Falcone, M., Kneisel, N., Hansson, J., Wang, W., Lutz, C., ... & Trumpp, A. (2017). BCAT1 restricts αKG levels in AML stem cells leading to IDH mut-like DNA hypermethylation. Nature, 551(7680), 384-388.
(\href{https://www.nature.com/articles/nature24294}{Nature})
}
\keyword{datasets}
