% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhosphoSitePlus.R
\docType{data}
\name{PhosphoSitePlus}
\alias{PhosphoSitePlus}
\title{PhosphoSitePlus}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(PhosphoSitePlus)
}
\description{
PhosphoSitePlus
}
\note{
Please see \href{https://www.phosphosite.org/staticDownloads}{PhosphoSitePlus (PSP)} for terms and conditions on the use PSP data.
WARNING: the PSP data is NOT for commercial use!
}
\references{
Hornbeck PV, Zhang B, Murray B, Kornhauser JM, Latham V, Skrzypek E PhosphoSitePlus, 2014: mutations, PTMs and recalibrations. Nucleic Acids Res. 2015 43:D512-20.
(\href{https://pubmed.ncbi.nlm.nih.gov/25514926/}{PubMed})
}
\keyword{datasets}
